김범현 매퍼, 서비스 user-admin 통합
