+++ ks44team04/src/main/java/ks44team04/admin/controller/AlertController.java
@@ -9,7 +9,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import ks44team04.admin.service.AlertService;
+import ks44team04.service.AlertService;
 import ks44team04.dto.Alert;
 import ks44team04.dto.AlertSend;
