+++ ks44team04/src/main/java/ks44team04/admin/controller/CouponController.java
@@ -11,7 +11,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import ks44team04.admin.service.CouponService;
+import ks44team04.service.CouponService;
 import ks44team04.dto.Coupon;
 
 @Controller
