--- ks44team04/src/main/java/ks44team04/admin/mapper/AlertMapper.java
@@ -1,19 +1,0 @@
-package ks44team04.admin.mapper;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-
-import ks44team04.dto.Alert;
-import ks44team04.dto.AlertSend;
-
-@Mapper
-public interface AlertMapper {
-	
-	//알림 전송내역 조회
-	public List<AlertSend> getAlertSendList();
-	
-	//알림 목록 조회
-	public List<Alert> getAlertList();
-	
-}
