--- ks44team04/src/main/java/ks44team04/admin/mapper/PaymentMapper.java
@@ -1,16 +1,0 @@
-package ks44team04.admin.mapper;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-
-import ks44team04.dto.PaymentTotal;
-
-@Mapper
-public interface PaymentMapper {
-	
-	public List<PaymentTotal> paymentDetailList();
-	
-	public List<PaymentTotal> paymentList();
-	
-}
