--- ks44team04/src/main/java/ks44team04/admin/mapper/ReportMapper.java
@@ -1,20 +1,0 @@
-package ks44team04.admin.mapper;
-
-import ks44team04.dto.Report;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface ReportMapper {
-
-
-    //신고 리스트
-    public List<Report> getReportList();
-    
-    //신고 하기 
-    public int report(Report report);
-    
-   
-}
