+++ ks44team04/src/main/java/ks44team04/mapper/PaymentMapper.java
@@ -1,0 +1,16 @@
+package ks44team04.mapper;
+
+import java.util.List;
+
+import org.apache.ibatis.annotations.Mapper;
+
+import ks44team04.dto.PaymentTotal;
+
+@Mapper
+public interface PaymentMapper {
+	
+	public List<PaymentTotal> paymentDetailList();
+	
+	public List<PaymentTotal> paymentList();
+	
+}
