+++ ks44team04/src/main/java/ks44team04/mapper/PointMapper.java
@@ -1,4 +1,4 @@
-package ks44team04.admin.mapper;
+package ks44team04.mapper;
 
 import java.util.List;
