+++ ks44team04/src/main/resources/mapper/AlertMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="ks44team04.admin.mapper.AlertMapper">
+<mapper namespace="ks44team04.mapper.AlertMapper">
 
 	<resultMap 	type="AlertSend"			id="AlertSendResultMap">
 		<result column="alert_send"			property="alertSend" />
@@ -11,7 +11,7 @@
 		<result column="send_id"			property="sendId" />
 		<result column="receive_date"		property="receiveDate" />
 		<result column="send_date"			property="sendDate" />
-			<association property="AlertInfo" 		javaType="AlertSend">
+			<association property="alertInfo" 		javaType="Alert">
 				<id 	column="alert_code"			property="alertCode" />
 				<result column="alert_content"		property="alertContent" />
 				<result column="reg_id"				property="regId" />
